Refactor block2
